+++ CHANGELOG.md
@@ -18,6 +18,7 @@
 - `datadogexporter`: Deprecate `service` setting in favor of `service.name` semantic convention (#8784)
 - `datadogexporter`: Deprecate `version` setting in favor of `service.version` semantic convention (#8784)
 - `datadogexporter`: Deprecate `GetHostTags` method from `TagsConfig` struct (#8975)
+- `jmxreceiver`: Deprecate JMX receiver ()
 
 ### 🚀 New components 🚀